+++ src/main/java/domain/DevelopmentCard.java
@@ -6,9 +6,9 @@ import java.util.List;
 public class DevelopmentCard extends Card {
     private final List<GemAmount> price;
     private final Crystal bonus;
-    private final byte level;
+    private final int level;
 
-    public DevelopmentCard(byte prestige, Crystal bonus, byte level) {
+    public DevelopmentCard(int prestige, Crystal bonus, int level) {
         super (prestige);
         this.price = new ArrayList<>();
         this.bonus = bonus;
